Provide an empty contended action instead of null
This contention handling is already active for exclusive locks, but
without an action it is ignoring requests from other parties.
With this, it sends confirmation that the request was received
and that lock releasing is in progress. Although the release itself
is not performed by the contended action, but manually after the
lock is no longer needed (which is the contract for exclusive locks).